feat: move AdditiveComplexFeedFilter and DefaultFeedOrder comparators to commons
Move DAL infrastructure files to commons module for KMP iOS migration:

- AdditiveComplexFeedFilter<T,U>: moved to commons, typealias at original
- DefaultFeedOrderEvent & DefaultFeedOrderCard: moved to commons alongside
  existing DefaultFeedOrder, re-exports at original location

Part of the KMP iOS migration (#2238).
